Move shared controls to a single point
this should make it easier to modify controls between projects.

also, fixed some excessive GC-ing on the extended progress bar. It will look funny in certain situations, but I don't know how else to handle it right now.
