+++ Controls/Enums/AllowedCharacters.cs
