+++ Controls/Enums/ProgressBarState.cs
