+++ Controls/Enums/TaskbarProgressState.cs
