--- youtube-dl-gui-updater/Controls/Natives/Consts.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace murrty.controls.natives;
-internal static class Consts {
-    internal const int WS_EX_COMPOSITED = 0x02000000;
-    internal const int WM_NCPAINT = 0x85;
-    internal const int WM_PAINT = 0xF;
-    internal const int WM_ERASEBKGND = 0x14;
-}
