--- youtube-dl-gui-updater/Controls/ProgressBarState.cs
@@ -1,22 +1,0 @@
-﻿namespace murrty.controls;
-
-using System;
-
-/// <summary>
-/// Represents the state of the progress bar control.
-/// </summary>
-[Flags]
-public enum ProgressBarState {
-    /// <summary>
-    /// Indicates the progress is succeeding so far, or finished.
-    /// </summary>
-    Normal = 0x1,
-    /// <summary>
-    /// Indicates the progress has encountered an error.
-    /// </summary>
-    Error = 0x2,
-    /// <summary>
-    /// Indicates the progress has been haulted. Can also represent a warning during progress.
-    /// </summary>
-    Paused = 0x3
-}
